Current objective :
    Create composite service for use 'module' into controller

#13 create a « add useCase »view

#12 create a « add usecase » controller
 * Creating Cscfa\Bundle\TwigUIBundle\Object\EnvironmentContainer
The EnvironmentContainer is used to store several controller environment variables.

 * Creating Cscfa\Bundle\TwigUIBundle\Object\ObjectsContainer
The ObjectsContainer is used to store several controller specific objects.

 * Creating Cscfa\Bundle\TwigUIBundle\Object\RenderRequest
The RenderRequest object is used to store a list of twig template to render and it's variables. The couple is stored as
TwigRequest.

 * Creating Cscfa\Bundle\TwigUIBundle\Object\TwigRequest\TwigRequest
The TwigRequest store a twig template path and it's variables.

TESTS
 * Creating test for class Cscfa\Bundle\TwigUIBundle\Object\ObjectsContainer
The ObjectsContainer is tested into Cscfa\Bundle\TwigUIBundle\Test\Objects\ObjectsContainerTest.

 * Creating test for class Cscfa\Bundle\TwigUIBundle\Object\TwigRequest\TwigRequest
The TwigRequest is tested into Cscfa\Bundle\TwigUIBundle\Test\Objects\TwigRequest\TwigRequestTest.